Moved atomic git functions to its own file, renamed logger
